+++ experimental/instrumentation-xhr/src/index.ts
@@ -1,1 +1,3 @@
 export { XHRInstrumentation } from './instrumentation';
+
+export { parseXHREvent } from './utils';